function/requirements.txt
adal==1.2.7
azure-common==1.1.28
azure-functions==1.11.2
azure-servicebus==0.50.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
distlib==0.3.4
filelock==3.7.0
idna==3.3
isodate==0.6.1
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.2.0
pipenv==2022.1.8
platformdirs==2.5.2
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==2.4.0
python-dateutil==2.8.2
python-http-client==3.3.7
requests==2.27.1
requests-oauthlib==1.3.1
sendgrid==6.3.0
six==1.16.0
uamqp==1.5.3
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
